requirements.txt
# Core ML Frameworks
torch>=2.1.0
transformers>=4.36.0
datasets>=2.15.0
peft>=0.7.1
trl>=0.7.4
bitsandbytes>=0.41.3
accelerate>=0.25.0

# Scientific & Utility
scipy>=1.11.0
numpy>=1.24.0
pyyaml>=6.0.1
tqdm>=4.66.0
packaging>=23.2

# Monitoring & Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
tensorboard>=2.15.1
wandb>=0.16.1

# UI (for HF Spaces)
gradio>=4.12.0
